# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "VSTR Management System API",
        "description": "Visitor Management System API Documentation",
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://localhost:8000/api",
            "description": "Local Server"
        },
        {
            "url": "https://api.vstrmgmt.com",
            "description": "Production Server"
        },
        {
            "url": "https://staging-api.vstrmgmt.com",
            "description": "Staging Server"
        }
    ],
    "paths": {
...
        "/super-admin/locations": {
            "get": {
                "tags": [
                    "Locations"
                ],
                "summary": "List locations",
                "description": "List locations (super-admin/admin)",
                "operationId": "listLocations",
                "responses": {
                    "200": {
                        "description": "Success"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            },
            "post": {
                "tags": [
                    "Locations"
                ],
                "summary": "Create location",
                "description": "Create a new location",
                "operationId": "createLocation",
                "responses": {
                    "200": {
                        "description": "Success"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
        "/super-admin/locations/list": {
            "get": {
                "tags": [
                    "Locations"
                ],
                "summary": "List simple locations",
                "description": "Return a simple list of locations",
                "operationId": "listSimpleLocations",
                "responses": {
                    "200": {
                        "description": "Success"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
        "/super-admin/locations/{location}": {
            "get": {
                "tags": [
                    "Locations"
                ],
                "summary": "Show location",
                "description": "Retrieve location details",
                "operationId": "showLocation",
                "parameters": [
                    {
                        "name": "location",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Success"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            },
            "put": {
                "tags": [
                    "Locations"
                ],
                "summary": "Update location",
                "description": "Update a location",
                "operationId": "updateLocation",
                "parameters": [
                    {
                        "name": "location",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Success"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            },
            "delete": {
                "tags": [
                    "Locations"
                ],
                "summary": "Delete location",
                "description": "Delete a location",
                "operationId": "deleteLocation",
                "parameters": [
                    {
                        "name": "location",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Success"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
        "/admin/locations/{location}/visits": {
            "get": {
                "tags": [
                    "Visits"
                ],
                "summary": "List visits (admin)",
                "description": "List visits for a location (admin)",
                "operationId": "adminListVisits",
                "parameters": [
                    {
                        "name": "location",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Success"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
...
    },
    "components": {
...
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "bearerFormat": "JWT",
                "scheme": "bearer"
            }
        }
    },
...
}